requirements.txt
streamlit>=1.28.0
PyMuPDF>=1.23.0
python-docx>=1.0.0
ebooklib>=0.18
beautifulsoup4>=4.12.0
lxml>=4.9.0
